Repository: Move package class into package.py
